get test_parameter_values passing after changes
